requirements.txt
llama-index-multi-modal-llms-gemini
llama-index-vector-stores-qdrant
llama-index-embeddings-gemini
llama-index-llms-gemini
llama-index==0.11.15
llama-index-readers-file
google-generativeai>=0.3.0 
matplotlib 
qdrant_client
python-dotenv
streamlit
llama-index-llms-nvidia
llama-index-embeddings-nvidia
markdownify
rapidocr-onnxruntime
pymupdf4llm
streamlit-authenticator
detectaicore
streamlit_pdf_viewer==0.0.12
pydantic==2.9.2
fastapi
ipython
orjson
